third_party/blink/web_tests/platform/mac/external/wpt/css/css-fonts/generic-family-keywords-003-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face matching for quoted and unquoted emoji (drawing text in a canvas)
  assert_equals: unquoted emoji does not match @font-face rule expected 25.0244140625 but got 125
[FAIL] @font-face matching for quoted and unquoted fangsong (drawing text in a canvas)
  assert_equals: unquoted fangsong does not match @font-face rule expected 25.0244140625 but got 125
[FAIL] @font-face matching for quoted and unquoted ui-serif (drawing text in a canvas)
  assert_equals: unquoted ui-serif does not match @font-face rule expected 25.0244140625 but got 125
[FAIL] @font-face matching for quoted and unquoted ui-sans-serif (drawing text in a canvas)
  assert_equals: unquoted ui-sans-serif does not match @font-face rule expected 25.0244140625 but got 125
[FAIL] @font-face matching for quoted and unquoted ui-monospace (drawing text in a canvas)
  assert_equals: unquoted ui-monospace does not match @font-face rule expected 25.0244140625 but got 125
[FAIL] @font-face matching for quoted and unquoted ui-rounded (drawing text in a canvas)
  assert_equals: unquoted ui-rounded does not match @font-face rule expected 25.0244140625 but got 125
Harness: the test ran to completion.
